Add docstrings and type hints
