+++ fresca_catalog/plot.py
@@ -1,112 +1,33 @@
-import pandas as pd
+"""Plotting utilities for the Fresca catalog."""
 import geopandas as gpd
-from shapely.geometry import Point
 from copy import deepcopy
 from typing import List
 
-from .catalog import Catalog, get_all_catalog_variables
+from .catalog import Catalog
+from .utils import build_agg_table
 
 import seaborn as sns
+import matplotlib.pyplot as plt
 from matplotlib.colors import LogNorm
-
-def build_agg_table(catalog: Catalog) -> pd.DataFrame:
-
-    base_cols = ['date', 'cruise_id', 'station', 'event_n', 'latitude', 'longitude']
-    entry_names = list(catalog.entries.keys())
-    dfs = []
-    for entry_name in entry_names:
-        df = catalog[entry_name].read()
-
-        cols_to_keep = deepcopy(base_cols)
-
-        # Handle CSV lat/lon columns
-        if all(c in catalog[entry_name].metadata for c in ['lon_col', 'lat_col']):
-            lon_col = catalog[entry_name].metadata['lon_col']
-            lat_col = catalog[entry_name].metadata['lat_col']
-        # Handle ERDDAP lat/lon columns (which have the unit appended)
-        else:
-            lon_col = next(c for c in df.columns if 'longitude' in c)
-            lat_col = next(c for c in df.columns if 'latitude' in c)
-        
-        # Rename lat/lon columns to standard names
-        df = df.rename(columns={lon_col: 'longitude', lat_col: 'latitude'})
-        lon_col = 'longitude'
-        lat_col = 'latitude'
-
-        if 'variables' not in catalog.metadata:
-            catalog.metadata['variables'] = get_all_catalog_variables(catalog)
-
-        for v in catalog.metadata['variables']:
-            if v in df.columns:
-                cols_to_keep.append(v)
-        cols_to_keep = list(set(cols_to_keep))
-        df = df[cols_to_keep]
-
-        df = df.reset_index(drop=True)
-        dfs.append(df)
-
-    df = pd.concat(dfs).fillna(0)
-
-    df['date'] = pd.to_datetime(df['date'])
-
-    if 'time_range' in catalog.metadata:
-        start, end = tuple(pd.to_datetime(d).tz_localize(None) for d in catalog.metadata['time_range'])
-    else:
-        start, end = df['date'].min(), df['date'].max()
-
-    df = df[df['date'].between(start, end)]
-
-    var_cols = df.columns.difference(base_cols)
-    import numpy as np
-    var_cols = [c for c in var_cols if np.issubdtype(df[c].dtype, np.number)]
-    df[var_cols] = (df[var_cols] > 0).astype(int)
-
-    # Collapse points down to per-station centroid
-    centroids = (
-        df.groupby('station')[[lon_col, lat_col]]
-        .mean()
-        .apply(lambda row: Point(row[lon_col], row[lat_col]), axis=1)
-    )
-    df.drop(columns=[lon_col, lat_col], inplace=True)
-
-    # Aggregate duplicate station visits per date
-    df = df.groupby(['station', 'date']).sum(numeric_only=True).reset_index()
-
-    date_range = pd.date_range(start, end)
-    stations = df['station'].unique()
-
-    full_index = pd.MultiIndex.from_product([stations, date_range], names=['station', 'date'])
-    df = df.set_index(['station', 'date'])
-    df = df.reindex(full_index, fill_value=0).reset_index()
-    df['event_n'] = 1
-        
-    df['geometry'] = df['station'].map(centroids)
-    gdf = gpd.GeoDataFrame(df)
-    gdf = gdf.set_crs(epsg=4326)
-    gdf = gdf[gdf.geometry.notnull()].copy()
-    gdf = gdf[gdf.geometry.apply(lambda g: g.is_valid if g else False)]
-
-    return gdf
-
 
 def plot_map(
     catalog: Catalog,
     metric: str = "mean",
     time_bin: str = "D",
     log: bool = False
-):
+) -> None:
     """Maps datasets in the catalog.
 
     Parameters
     ----------
     catalog : Catalog
         The catalog containing the datasets to map.
-    variables : list
-        The variables map.
+    metric : str
+        The metric to plot. Either 'mean' or 'sum'. Defaults to 'mean'.
+    time_bin : str
+        The time bin to aggregate data by. Defaults to 'D' (daily). Uses pandas offset alias strings.
     log : bool
         Whether or not to log normalize counts. Defaults to False.
-    hex : bool
-        Whether or not to aggregate observations into hexes. Defaults to True.
     """
     base_cols= ['date', 'station', 'event_n', 'geometry']
 
@@ -153,23 +74,25 @@ def plot_map(
 
 def plot_grid(
     catalog: Catalog,
-    metric: str = 'mean',
     stations: List[str] = None,
-    log: bool = False,
-    time_bin: str = "D"
-):
+    metric: str = 'mean',
+    time_bin: str = "D",
+    log: bool = False
+) -> None:
     """Maps datasets in the catalog.
 
     Parameters
     ----------
     catalog : Catalog
         The catalog containing the datasets to map.
-    variables : list
-        The variables map.
+    stations : list
+        The stations to plot. If None, plots all stations. Defaults to None.
+    metric : str
+        The metric to plot. Either 'mean' or 'sum'. Defaults to 'mean'.
+    time_bin : str
+        The time bin to aggregate data by. Defaults to 'D' (daily). Uses pandas offset alias strings.
     log : bool
         Whether or not to log normalize counts. Defaults to False.
-    hex : bool
-        Whether or not to aggregate observations into hexes. Defaults to True.
     """
     if log:
         norm = LogNorm()
@@ -221,24 +144,25 @@ def plot_grid(
 
 def plot_timeseries(
     catalog: Catalog,
-    metric: str = 'mean',
     stations: List[str] = None,
-    variables: List[str] = None,
-    log: bool = False,
-    time_bin: str = "D"
-):
+    metric: str = 'mean',
+    time_bin: str = "D",
+    log: bool = False
+) -> None:
     """Maps datasets in the catalog.
 
     Parameters
     ----------
     catalog : Catalog
         The catalog containing the datasets to map.
-    variables : list
-        The variables map.
+    stations : list
+        The stations to plot. If None, plots all stations. Defaults to None.
+    metric : str
+        The metric to plot. Either 'mean' or 'sum'. Defaults to 'mean'.
+    time_bin : str
+        The time bin to aggregate data by. Defaults to 'D' (daily). Uses pandas offset alias strings.
     log : bool
         Whether or not to log normalize counts. Defaults to False.
-    hex : bool
-        Whether or not to aggregate observations into hexes. Defaults to True.
     """
     base_cols= ['date', 'station', 'event_n']
 
@@ -286,7 +210,6 @@ def plot_timeseries(
             for col in agg_table.columns:
                 agg_table[col] = agg_table[col] / len(agg_table.columns)
 
-    import matplotlib.pyplot as plt
     ax = agg_table.plot(kind='bar', stacked=True, logy=log, width=1.0)
     xticks = range(0, len(agg_table), max(1, len(agg_table)//10))  # adjust tick frequency
     ax.set_xticks(xticks)
